Use tgmath.h so that we can use fabs without warnings on both 32- and 64-bit builds.

+++ ZSSRichTextEditor/Third Party/CYRTextView.m
@@ -34,6 +34,8 @@
 // IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+#import <tgmath.h>
+
 #import "CYRTextView.h"
 #import "CYRLayoutManager.h"
 #import "CYRTextStorage.h"
@@ -229,7 +231,7 @@ static const float kCursorVelocity = 1.0f/8.0f;
     if (gestureRecognizer == _singleFingerPanRecognizer || gestureRecognizer == _doubleFingerPanRecognizer)
     {
         CGPoint translation = [gestureRecognizer translationInView:self];
-        return fabsf(translation.x) > fabsf(translation.y);
+        return fabs(translation.x) > fabs(translation.y);
     }
     
     return YES;
@@ -261,11 +263,11 @@ static const float kCursorVelocity = 1.0f/8.0f;
     
     if (cursorLocation > startRange.location)
     {
-        self.selectedRange = NSMakeRange(startRange.location, fabsf(startRange.location - cursorLocation));
+        self.selectedRange = NSMakeRange(startRange.location, fabs(startRange.location - cursorLocation));
     }
     else
     {
-        self.selectedRange = NSMakeRange(cursorLocation, fabsf(startRange.location - cursorLocation));
+        self.selectedRange = NSMakeRange(cursorLocation, fabs(startRange.location - cursorLocation));
     }
 }
